# internal/transport/rest/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for SG Emulator with cryptographically signed requests",
        "title": "SG Emulator REST API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/accounts/me": {
            "post": {
                "description": "Get account details including balance and transaction history. Requires cryptographically signed request with Ed25519 signature and X.509 certificate.\n\nThe request must be wrapped in a SignedEnvelope structure containing:\n- payload: The AccountRequest with your account_id\n- signature: Ed25519 signature with algorithm, value (base64-encoded), signer_id, and timestamp\n- certificate: PEM-encoded X.509 certificate containing your public key\n\nThe signature must be created by signing the JSON bytes of the payload using your account's Ed25519 private key.\nThe signer_id in the signature must match the account_id in the payload.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "Get your own account details",
                "parameters": [
                    {
                        "description": "Signed account request with Ed25519 signature",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rest.SignedAccountRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.AccountResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body or missing signature",
                        "schema": {
                            "$ref": "#/definitions/rest.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Attempting to access another account",
                        "schema": {
                            "$ref": "#/definitions/rest.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Account not found",
                        "schema": {
                            "$ref": "#/definitions/rest.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Get the health status of the REST API",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.HealthResponse"
                        }
                    }
                }
            }
        },
        "/transfer": {
            "post": {
                "description": "Transfer funds from one account to another. Requires cryptographically signed request with Ed25519 signature.\n\nThe request must be wrapped in a SignedEnvelope containing:\n- payload: TransferRequest with from (sender account ID), to (recipient account ID), amount, nonce (random 16-byte hex string), and timestamp\n- signature: Ed25519 signature with algorithm, value (base64-encoded), signer_id (must match 'from'), and timestamp\n- certificate: PEM-encoded X.509 certificate of the sender\n\nThe nonce prevents replay attacks. Each transfer must use a unique nonce.\nThe signature is created by signing the canonical JSON bytes of the TransferRequest payload.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transfer"
                ],
                "summary": "Transfer funds between accounts",
                "parameters": [
                    {
                        "description": "Signed transfer request with Ed25519 signature",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rest.SignedTransferRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.TransferResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request, missing fields, or transfer failed",
                        "schema": {
                            "$ref": "#/definitions/rest.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "crypto.Signature": {
            "type": "object",
            "properties": {
                "algorithm": {
                    "description": "\"Ed25519\"",
                    "type": "string"
                },
                "signer_id": {
                    "description": "Account ID (hash of public key)",
                    "type": "string"
                },
                "timestamp": {
                    "description": "Unix timestamp when signed",
                    "type": "integer"
                },
                "value": {
                    "description": "Raw signature bytes (64 bytes for Ed25519)",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "crypto.SignedEnvelope-crypto_TransferRequest": {
            "type": "object",
            "properties": {
                "certificate": {
                    "description": "PEM-encoded X.509 certificate",
                    "type": "string"
                },
                "payload": {
                    "$ref": "#/definitions/crypto.TransferRequest"
                },
                "signature": {
                    "$ref": "#/definitions/crypto.Signature"
                }
            }
        },
        "crypto.SignedEnvelope-rest_AccountRequest": {
            "type": "object",
            "properties": {
                "certificate": {
                    "description": "PEM-encoded X.509 certificate",
                    "type": "string"
                },
                "payload": {
                    "$ref": "#/definitions/rest.AccountRequest"
                },
                "signature": {
                    "$ref": "#/definitions/crypto.Signature"
                }
            }
        },
        "crypto.TransferRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "from": {
                    "type": "string"
                },
                "nonce": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "integer"
                },
                "to": {
                    "type": "string"
                }
            }
        },
        "rest.AccountRequest": {
            "type": "object",
            "properties": {
                "account_id": {
                    "type": "string",
                    "example": "6c439a07c32f7fb09c29403d8d2e4e47b8c5e8a9"
                }
            }
        },
        "rest.AccountResponse": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "number",
                    "example": 100.5
                },
                "id": {
                    "type": "string",
                    "example": "6c439a07c32f7fb09c29403d8d2e4e47b8c5e8a9"
                },
                "transactions": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "additionalProperties": true
                    }
                }
            }
        },
        "rest.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Invalid request"
                }
            }
        },
        "rest.HealthResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "ok"
                },
                "type": {
                    "type": "string",
                    "example": "rest"
                }
            }
        },
        "rest.SignedAccountRequest": {
            "type": "object",
            "properties": {
                "signed_envelope": {
                    "$ref": "#/definitions/crypto.SignedEnvelope-rest_AccountRequest"
                }
            }
        },
        "rest.SignedTransferRequest": {
            "type": "object",
            "properties": {
                "signed_envelope": {
                    "$ref": "#/definitions/crypto.SignedEnvelope-crypto_TransferRequest"
                }
            }
        },
        "rest.TransferResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        }
    }
}